+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.ReservationTimeCreationRequest;
-import roomescape.dto.ReservationTimeWithBookState;
+import roomescape.dto.request.ReservationTimeCreationRequest;
+import roomescape.dto.other.ReservationTimeWithBookState;
 import roomescape.service.ReservationTimeService;
 
 @RestController
