+++ src/main/java/roomescape/controller/ThemeController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Theme;
-import roomescape.dto.ThemeCreationRequest;
+import roomescape.dto.request.ThemeCreationRequest;
 import roomescape.service.ThemeService;
 
 @RestController
