+++ src/main/java/roomescape/dto/request/ReservationTimeCreationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 import java.time.LocalTime;
