+++ src/main/java/roomescape/dto/request/ThemeCreationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 public record ThemeCreationRequest(String name, String description, String thumbnail) {
