+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -6,8 +6,8 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.ReservationTimeCreationRequest;
-import roomescape.dto.ReservationTimeWithBookState;
+import roomescape.dto.request.ReservationTimeCreationRequest;
+import roomescape.dto.other.ReservationTimeWithBookState;
 import roomescape.exception.BadRequestException;
 import roomescape.exception.NotFoundException;
 import roomescape.repository.ReservationRepository;
