+++ src/test/java/roomescape/dto/ReservationTimeCreationRequestTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.dto.request.ReservationTimeCreationRequest;
 
 class ReservationTimeCreationRequestTest {
