+++ src/test/java/roomescape/dto/ThemeCreationRequestTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import roomescape.dto.request.ThemeCreationRequest;
 
 class ThemeCreationRequestTest {
